+++ test/legacy/41-device-api-v1.spec.ts
@@ -667,10 +667,12 @@ describe('SupervisorAPI [V1 Endpoints]', () => {
 	describe('POST /v1/update', () => {
 		let configStub: SinonStub;
 		let targetUpdateSpy: SinonSpy;
+		let readyForUpdatesStub: SinonStub;
 
 		before(() => {
 			configStub = stub(config, 'get');
 			targetUpdateSpy = spy(TargetState, 'update');
+			readyForUpdatesStub = stub(apiBinder, 'isReadyForUpdates').returns(true);
 		});
 
 		afterEach(() => {
@@ -680,6 +682,7 @@ describe('SupervisorAPI [V1 Endpoints]', () => {
 		after(() => {
 			configStub.restore();
 			targetUpdateSpy.restore();
+			readyForUpdatesStub.restore();
 		});
 
 		it('returns 204 with no parameters', async () => {
